# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "myrti",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/assets/timeline": {
      "get": {
        "tags": [
          "crate::routes::timeline"
        ],
        "operationId": "getTimeline",
        "parameters": [
          {
            "name": "lastAssetId",
            "in": "query",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/AssetId"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "maxCount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "lastFetch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineChunk"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/timeline/sections/{id}": {
      "get": {
        "tags": [
          "crate::routes::timeline"
        ],
        "operationId": "getTimelineSegments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Section id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineSegmentsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Asset": {
        "type": "object",
        "required": [
          "assetId",
          "takenDate",
          "repFile"
        ],
        "properties": {
          "assetId": {
            "$ref": "#/components/schemas/AssetId"
          },
          "repFile": {
            "$ref": "#/components/schemas/AssetFile"
          },
          "takenDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AssetFile": {
        "type": "object",
        "required": [
          "fileId",
          "assetRootId",
          "pathInRoot",
          "width",
          "height",
          "addedAt",
          "rotationCorrection",
          "mirrorCorrection",
          "mimeType"
        ],
        "properties": {
          "addedAt": {
            "type": "string",
            "format": "date-time"
          },
          "assetRootId": {
            "$ref": "#/components/schemas/AssetRootDirId"
          },
          "fileId": {
            "$ref": "#/components/schemas/FileId"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "mimeType": {
            "type": "string"
          },
          "mirrorCorrection": {
            "$ref": "#/components/schemas/MirrorCorrection"
          },
          "pathInRoot": {
            "type": "string"
          },
          "rotationCorrection": {
            "type": "integer",
            "format": "int32"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AssetId": {
        "type": "string"
      },
      "AssetRootDirId": {
        "type": "string"
      },
...
      "AssetSeriesId": {
        "type": "string"
      },
      "AssetSpe": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/Image"
              },
              {
                "type": "object",
                "required": [
                  "assetType"
                ],
                "properties": {
                  "assetType": {
                    "type": "string",
                    "enum": [
                      "image"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/Video"
              },
              {
                "type": "object",
                "required": [
                  "assetType"
                ],
                "properties": {
                  "assetType": {
                    "type": "string",
                    "enum": [
                      "video"
                    ]
                  }
                }
              }
            ]
          }
        ],
        "discriminator": {
          "propertyName": "assetType"
        }
      },
...
      "AssetWithSpe": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Asset"
          },
          {
            "$ref": "#/components/schemas/AssetSpe"
          },
          {
            "type": "object"
          }
        ]
      },
...
      "FileId": {
        "type": "string"
      },
...
      "Image": {
        "type": "object",
        "required": [
          "representations"
        ],
        "properties": {
          "representations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageRepresentation"
            }
          }
        }
      },
      "ImageRepresentation": {
        "type": "object",
        "required": [
          "id",
          "format",
          "width",
          "height",
          "size"
        ],
        "properties": {
          "format": {
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "MirrorCorrection": {
        "type": "string",
        "enum": [
          "None",
          "Horizontal",
          "Vertical"
        ]
      },
      "SegmentType": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "start",
              "end",
              "type"
            ],
            "properties": {
              "end": {
                "type": "string",
                "format": "date-time"
              },
              "start": {
                "type": "string",
                "format": "date-time"
              },
              "type": {
                "type": "string",
                "enum": [
                  "dateRange"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "id",
              "type"
            ],
            "properties": {
              "id": {
                "$ref": "#/components/schemas/TimelineGroupId"
              },
              "name": {
                "type": "string",
                "nullable": true
              },
              "type": {
                "type": "string",
                "enum": [
                  "userGroup"
                ]
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "type"
        }
      },
...
      "TimelineChunk": {
        "type": "object",
        "description": "Response for a request for the next part of the timeline to display\n\n`groups` are always whole, not sliced in the middle. Either TimelineGroup or Day\n`date` is the date before queries are made",
        "required": [
          "date",
          "changedSinceLastFetch",
          "groups"
        ],
        "properties": {
          "changedSinceLastFetch": {
            "type": "boolean"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineGroup"
            }
          }
        }
      },
      "TimelineGroup": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TimelineGroupType"
          },
          {
            "type": "object",
            "required": [
              "assets"
            ],
            "properties": {
              "assets": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AssetWithSpe"
                }
              }
            }
          }
        ]
      },
      "TimelineGroupId": {
        "type": "string"
      },
      "TimelineGroupType": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "date",
              "type"
            ],
            "properties": {
              "date": {
                "type": "string",
                "format": "date"
              },
              "type": {
                "type": "string",
                "enum": [
                  "day"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "groupTitle",
              "groupStartDate",
              "groupEndDate",
              "groupId",
              "type"
            ],
            "properties": {
              "groupEndDate": {
                "type": "string",
                "format": "date-time"
              },
              "groupId": {
                "type": "string"
              },
              "groupStartDate": {
                "type": "string",
                "format": "date-time"
              },
              "groupTitle": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "group"
                ]
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "type"
        }
      },
      "TimelineItem": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/AssetWithSpe"
              },
              {
                "type": "object",
                "required": [
                  "itemType"
                ],
                "properties": {
                  "itemType": {
                    "type": "string",
                    "enum": [
                      "asset"
                    ]
                  }
                }
              }
            ]
          },
          {
            "type": "object",
            "required": [
              "seriesId",
              "assets",
              "selectionIndices",
              "totalSize",
              "itemType"
            ],
            "properties": {
              "assets": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AssetWithSpe"
                },
                "description": "assets[0] is most recent, last is oldest asset"
              },
              "itemType": {
                "type": "string",
                "enum": [
                  "assetSeries"
                ]
              },
              "selectionIndices": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "minimum": 0
                }
              },
              "seriesId": {
                "$ref": "#/components/schemas/AssetSeriesId"
              },
              "totalSize": {
                "type": "integer",
                "minimum": 0
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "itemType"
        }
      },
...
      "TimelineSegment": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SegmentType"
          },
          {
            "type": "object",
            "required": [
              "sortDate",
              "items"
            ],
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TimelineItem"
                }
              },
              "sortDate": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "TimelineSegmentsResponse": {
        "type": "object",
        "required": [
          "segments"
        ],
        "properties": {
          "segments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineSegment"
            }
          }
        }
      },
      "Video": {
        "type": "object",
        "required": [
          "hasDash"
        ],
        "properties": {
          "hasDash": {
            "type": "boolean"
          }
        }
      }
    }
  },
...
}